Bug fixes for Farmer List Screen
