+++ app/Orchid/Layouts/Farmer/FarmerListLayout.php
@@ -2,7 +2,7 @@
 
 namespace App\Orchid\Layouts\Farmer;
 
-use App\Models\Farmer\Farmer_profile;
+use App\Models\Farmer_profile;
 use Orchid\Screen\TD;
 use Orchid\Screen\Actions\Link;
 use Orchid\Screen\Layouts\Table;
